Home/Index, DataBase sorunları fix
Indexte ai tarafından oluşturulmuş ve düzenlenmiş bir tasarım bıraktım.
Database'deki telefon numarasını kaydetmeme ve migrations tarafından telefon no oluşturulmamaması fixlendi.

+++ WebProje/WebProje/Models/Kullanici.cs
@@ -11,7 +11,7 @@ namespace WebProje.Models
         public string Email { get; set; } = "";
         public string Sifre { get; set; } = "";
         public string Rol { get; set; } = "User"; // Admin veya Kullanici
-        public int TelNo { get; set; } 
+        public long TelNo { get; set; } 
 
         public List<Siparis> Siparisler { get; set; }
     }
